refactor: extract CLI definitions from main.rs into cli module
